webapp/requirements.txt
Flask-Babel==1.0.0
Flask-SSLify==0.1.5
Flask-SocketIO==5.3.3
SQLAlchemy==1.4.46
Werkzeug==2.2.3
alembic==1.10.3
eventlet==0.33.3
flask-sqlalchemy==3.0.3
flask_limiter==3.3.0
hypercorn==0.14.3
itsdangerous==2.1.2
oauthlib==3.2.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
quart==0.18.3
quart-sqlalchemy==2.0.0
redis[hiredis]==4.5.4
requests==2.28.2
requests_oauthlib==1.3.1
